Disable prepared_statement test
I could not get it to work while bootstrapping the docker images and
then running the test suite (it works fine using the Docker image
published on Docker Hub).

I discussed with others, and with prepared_statements marked
experimental, we will disable the tests for now so that the 4.0 branch
can start based on a passing testsuite.

Signed-off-by: Morgan Tocker <tocker@gmail.com>