cluster: deflake worker test
There is a race where PRS finishes so fast that the worker doesn't
catch the error from the read-only master. So, looking for that
failure makes the worker test flaky. Somehow, this ISM improvement
seems to make this test more flakly.

In any case, correctness is verified at the end by comparing rows.
So, there's not need to count the number of errors encountered
by vtworker.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>