Parse EOF-flavored OK packets like a 5.8+ client (#571)
* Parse EOF-flavored OK packets like a 5.8+ client

Pure EOF packets are deprecated as of MySQL 5.7.5.  Clients aware of this
fact now advertise a capability, `CLIENT_DEPRECATE_EOF`, and recent
servers echo it back.  If both endpoints have that capability, then
EOF-flavored OK packets are sent instead of pure EOF packets to mark the
end of a result set.

An OK packet (whether EOF-flavored or not) can be arbitrarily large,
because it has variable-length strings in it.  So the old implementation
of `isEOFPacket`, which can only return true for lengths <= 8 bytes, is no
longer valid.  The official mysql client library treats a packet beginning
with 0xfe (the EOF marker) as an EOF-flavored OK packet if the length is
less than 16MiB, a check sufficient to rule out variable-length integers
prefixing length-encoded strings, which is the other way that the MySQL
protocol uses 0xfe.

So that's what we do now.  If `CLIENT_DEPRECATE_EOF` is set, an EOF can be
up to 16 MiB.  If not, it's still limited to <= 8 bytes.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Add a test

Signed-off-by: Patrick Reynolds <patrick@piki.org>