Add a sqlcommenter parser for Insights (#573)
* Add a sqlcommenter parser

The `SQL` field in `LogStats` contains normalized SQL, but with
`/* ... */` comments still in it.  That messes up our aggregation, so we
need to split those comments off.

We also now parse the split-off comments for sqlcommenter tags and put
those tags in Kafka `Query` messages.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Add a test

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Add function comments

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* WriteByte is cheaper than WriteRune

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Harden sqlcommenter parsing, support \escapes

This fixes potential index-out-of-range crashes if people pass us
SQL comments that aren't in valid sqlcommenter format.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* One more test

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Add more comments-parsing tests

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Another test

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Add some invariants

Parsers are hard.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Moar tests

Signed-off-by: Patrick Reynolds <patrick@piki.org>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>