Limit some Insights log messages to once per 15s (#624)
* Limit some log messages to once per 15s

These two log messages are under the client's control: exceeding the
maximum number of patterns to aggregate, and exceeding the Kafka buffer
size.  A very busy client may (intentionally or not) exceed these limits,
and we don't want one log line per statement when that happens.

So we keep a counter of overflows and log about it only at the end of a
15s interval.

Other chatty log messages that aren't under the client's control are left
intact here.  In particular, Kafka errors, type-cast errors, and
serialization errors are things where we'll want to see `err.Error()`, so
we don't just roll them up as a uint.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Fix `log.Infof` arguments

Signed-off-by: Patrick Reynolds <patrick@piki.org>