Log `RowsRead`, `RowsAffected`, and `RowsReturned` for streaming queries (#693)
* Log rows returned/affected in StreamExecute

* Plumb rows-read through streaming results

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Return `packetOK.info` for streaming queries

Streaming queries call `FetchNext` repeatedly to keep fetching rows, until
it returns `nil` for the row, which signifies EOF.  But we want the `Info`
field that comes along with the EOF, because it has the `rows_read`
metric.

So I've added a third return value, a `*sqltypes.Result`, to return from
`FetchNext`.  When we get an EOF-flavored OK packet and the caller wants
`Info`, we parse the packet and return a `sqltypes.Result` containing the
correct `Info`, along with a nil row.

Co-authored-by: Rafer Hazen <rafer@ralua.com>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>