Structured schemadiff errors (#10356) (#708)
* Move error definitions to separate file

This starts the refactoring for more structured errors by moving it into
a separate file first.

Also removes a few unused error types.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Update single usage of github.com/pkg/errors to errors

This moves the single usage of the github.com/pkg/errors package to use
regular errors.

The rest is all in schemadiff and will be cleaned up with structured
errors. It removes the package as a direct dependency (but still other
Vitess dependencies pull it in).

The usage here does not need the package features at all.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Move column level errors to use structured errors

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Move all apply style errors to structured errors

This moves over all the Apply style errors to be using a structured
version of the error instead.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* More structured errors for parsing failures

This adds a few more structured errors, for example for when we fail to
parse things or handle unknown table options.

Also adds a few more table options the parser knows about but that we
can ignore otherwise but don't want to error on.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Always check complete parsing before other operations

It was still possible to use `NewSchemaFromStatements` to trigger a
panic since it would try to normalize without checking if the statement
was fully parsed.

This moves the check into the helper function used for create new
statements for schemadiff and always ensures we check parse status
before we try to normalize.

Also parse everywhere including in tests with strict mode to show errors
as soon as possible.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Move all to regular errors package

Now that all specific usage for github.com/pkg/errors is removed, let's
remove the package usage and only use regular errors. We don't need it
anymore.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>