Limit the number of individual queries reported (#694)
* Change default rows-read threshold to 10000

`planetscale/planetscale` has a feature-flag query that reads 2,663 rows
per query, and returns 2,616 of them.  It also runs more than a million
times per day.  That made us reconsider the rows-read threshold -- 1000
rows read is bad, but maybe not "always report this statement
individually" bad.

So now the threshold is 10000.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Report at most 100 queries per 15s

`planetscale/planetscale` has a feature-flag query that is interesting
over a million times per day.  That's a lot of data to store in the
`branch_query` table, and far more than would actually be useful to an end
user.

The number of individual queries we report is currently unlimited -- every
error, everything over 1s, and everything over 10000 rows read.

To protect Insights from corner cases, let's put a cap on that.  The cap
is now 100 queries per vtgate per 15s.  For a cluster with six vtgates, as
`planetscale/planetscale` has, that's still 3.5M rows per day.  But at
least it's finite.

It's also configurable, so we can tune it quickly and on a per-cluster
basis, if needed, with a PR to Singularity.

Signed-off-by: Patrick Reynolds <patrick@piki.org>