Track whether or not ls.SQL is normalized (#711)
Previously, Insights used a regular expression on `ls.ErrorStr()` to try
to guess, based on the contents of the error, if it could assume that
`ls.SQL` was normalized.

But of course, vtgate knows if `ls.SQL` is normalized.  So let's just
track that explicitly.

That gets Insights out of the regex-based guessing game, and means we can
report `ls.SQL` to Kafka in some cases where it would've been unclear from
the error text if that was a safe thing to do.  In particular,
`supported only for primary tablet type` errors will now come with the
actual SQL that couldn't be executed.

Signed-off-by: Patrick Reynolds <patrick@piki.org>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>