Send unnormalized queries to Insights, if requested (#813)
* Add `-insights_raw_queries` command-line option

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* WIP - capture raw SQL

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Update psevents dependency to get raw-sql fields

https: //github.com/planetscale/psevents/compare/238f3799dfa0..98d881c120bf
Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Size-limit raw SQL, add unit tests

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Make insights_raw_queries_max_length configurable

It was hard-coded 4096 bytes, now it's a command-line option defaulting to
8192 bytes.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Truncate UTF-8 strings safely

A string with multibyte characters in it must be cut at a character
boundary.  The limit is still on the number of bytes in the resulting
string, but the resulting string might now be up to 3 characters shorter
than the limit, if the boundary character was a 4-byte UTF-8 sequence.

String cutting is O(n), where n is the byte-size limit, but only happens
for strings that are over the limit to begin with, so this is an
acceptable cost.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Truncate UTF-8 strings in O(1) time

Signed-off-by: Patrick Reynolds <patrick@piki.org>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>