Cherry pick topo improvements (#10906) (#916)
* Add List() implementation for memory topo

This adds a List() implementation for the memory topo and adds a bunch
of tests for this function as well. These tests are only run when the
implementation indicates proper support for List.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Refactor Watch() to use context properly

A function that can spawn a goroutine should use the passed in context
for that and not return a cancel function itself. That is considered an
anti-pattern, so let's change this all to more idiomatic Go patterns.

This also ensures we can now set a proper timeout on the initial data
retrieval for a watch using the configured topo timeout.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Add recursive watcher for topo

This adds a new function to the topo interface which allows for
recursive watching on the topo.

Recursive watching can be used to significantly improve performance in a
few cases where we want to monitor changes across a prefix.

There's no immediate usage added yet, that will be done in follow up
changes.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Don't allow expired contexts to retrieve topo connection

Once a context is expired, we shouldn't return a valid cell anymore,
even if it's the global cell.

Local cells would fail in this case as well, so this ensures we match
that behavior.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Add WaitForNewLeader to election API

This allows for more efficiently waiting for a new leader to be elected.
It also allows for a continued stream of changes during the election
lifetime.

This is very useful for example when something else needs to be
triggered on an election change and we don't need to use a busy wait
loop for it in that case.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Update entries also when receiving direct errors

Also when we get an error directly from the watcher on the initial
entry, we want to update the tracking value with that error. This is
needed to ensure we handle errors like a file not existing and appearing
later properly.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Fix broken k8stopo List implementation

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Add back still used code

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Fix import

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>