Add `user_supplied_id` workaround to Nightly Build GitHub workflow (#1054)
* Add `user_supplied_id` workaround to Nightly Build GitHub workflow

We're going to invoke this Workflow from Temporal (from
https://github.com/planetscale/infra-temporal-workflows specifically) as
part of automating Vitess releases and we need some way to retrieve the
workflow that was started by the workflow dispatch event.

GitHub doesn't have a built in way to do that (it doesn't return a
correlation ID or similar when you create the dispatch event) so we need
to inject our own ID that we can use to look up the workflow run later.

* Add `user_supplied_id` workaround to Copy to main workflow

We'll also need it here, since we'll be invoking it from Temporal
workflows.

Signed-off-by: Rohit Nayak <rohit@planetscale.com>