Add default lower stack limit (#11569) (#1212)
This changes the default maximum stack size we allow to 64MiB. There
should not be any algorithms in Vitess or logic that ever needs this
kind of stack size and it's a really high safe limit we believe for now.

It also adds a knob to tune in case the limit doesn't suffice, or if
someone wants to tighten it further.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>