boost: Order By (#1322)
* [wip] order by in the client

Signed-off-by: Vicent Marti <vmg@strn.cat>

* Revert "error out if ORDER BY is at the outer SELECT or UNION (#1313)"

This reverts commit 05a92acbc1acf86a0bce720abc1b03bf2fc0b6b1.

* integration: better lookup helpers

Signed-off-by: Vicent Marti <vmg@strn.cat>

* integration: results!

Signed-off-by: Vicent Marti <vmg@strn.cat>

* Fix ordering of TopK

It looks like even in Noria, ordering of TopK results is wrong. While we
do originally order results in the correct ordering in the `current`
slice, it is processed in a way that essentially reverses the order.

The issue is that larger values are seen as the top values, but that's
not how `ORDER BY` functions. `ORDER BY` has a default ordering of
ascending, so any lower value is the default to retain. A descending
order is explicit with `ORDER BY DESC`.

This changes the processing of the current list and slightly changes the
loop calls to ensure we deal with shorter than `k` slices as well. It
also updates the tests accordingly because those were reversed as well.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* integration: fix topk tests

Signed-off-by: Vicent Marti <vmg@strn.cat>

* boost: update sizegen

Signed-off-by: Vicent Marti <vmg@strn.cat>

Signed-off-by: Vicent Marti <vmg@strn.cat>
Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>
Co-authored-by: Dirkjan Bussink <d.bussink@gmail.com>