Update all the Go dependencies (#11741) (#1348)
* Update all the Go dependencies (#11741)

* Update all the Go dependencies

This seems to work for the basics just fine, so let's have CI take a run
at this as well to update these.

Only one small update to the Azure blob storage handling seems needed so
far.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Use correct proto comparisons

We're using `reflect.DeepEqual` or `testify` helpers that essentially
use that in a number of places are comparing protobufs. This is not
supported though, protobufs are not comparable with `reflect.DeepEqual`.

This is exposed because of the tiny patch bump of protobuf which changes
some internal optimization of how it initializes protobufs that breaks
all this.

Instead, move to the appropriate helpers here.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Update generated proto

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Keep btree at older version

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>
Signed-off-by: Rohit Nayak <rohit@planetscale.com>