Fix parsing and normalization of floating point types (#12009) (#1454)
* Fix parsing and normalization of floating point types (#12009)

* Fix parsing of float type with single length attribute

MySQL can also parse this so ensure that we're compatible here. This is
deprecated but it can still be used to create tables but in MySQL 8 this
length is ignored on the final table.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Add float4 and float8 aliases for float and double

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Normalize floating point values according to MySQL rules

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Simplify parser with reuse

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* Keywords must be unique and can't be reused (#12044)

In https://github.com/vitessio/vitess/pull/12009 I added aliases for
float4 / float8 but that was done wrongly. We can't alias these with the
same keyword, or we break functions like SQLTypeToQueryType because we
only have one version of it in the keywords list.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>