evalengine: Add basic support for information functions
This adds support for handling `USER()`, `DATABASE()` and `VERSION()`.
The last two have fairly well defined semantics, but the first is really
a best effort.

`USER()` returns the effective caller id if it's set, and otherwise
falls back to the default `vt_app@localhost` that Vitess uses. Today the
value that is returned isn't very useful on Vitess anyway since it
returns the MySQL level user. So the called id is in practice more
useful already if available.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>