Suppress comments in Insights output (#2017)
* Omit comments from Insights normalized SQL

Reusing the parsed AST in #1796 led to Insights emitting normalized SQL
statements with comments in them.  That's most noticeable when vtgate puts
a `SET_VAR(...)` comment right after the first keyword of the query.  The
Insights web page truncates queries after two wrapped lines, so often all
a user sees is `select /*+ SET_VAR(...` and none of the actual query logic
that comes after it.

So now we emit normalized statements that do not contain any of the
comments.  All comments are omitted, whether they are structured
sqlcommenter tag comments, freeform comments, or vtgate internal `SET_VAR`
comments.

This change has no effect on parsing sqlcommenter tags, and no effect on
other Vitess-internal users of the `SET_VAR` comments.  All it changes is
what Insights sends to Kafka.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Return only comments, not SQL, in extractComments

Insights parses ls.SQL looking for comments, but it no longer needs the
version of the string with the comments removed.  All it needs are the
actual comments.  So the `extractComments` function can save itself the
cost of some string copies and return only the comments.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

---------

Signed-off-by: Patrick Reynolds <patrick@piki.org>