Insights: normalize unary operators (#2016)
* Normalize unary operators before `?`

Currently, unary operators aren't normalized, so a query like
  select * from users where id=-1
normalizes to
  select * from users where id = -?

That's confusing to look at, it messes up the parser in api-bb, and it can
lead to 2^n blowout if several bound values can be either positive or
negative.

So, now we combine any unary operator (+, -, !, ~, or N) into the
expression that follows it, iff the expression would normalize to a `?`.
So the query above becomes
  select * from users where id = ?

Unary operators that precede more complex expressions aren't elided.  For
example,
  select * from users where id = -balance
  select * from users where id = +(select id from users limit 1)
retain their unary operators, because the expression after the unary isn't
a literal.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Squash introducer expressions like unary operators

`_utf8'foo'` should normalize to `?`

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* DRY the list of node types

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Remove now-redundant case statement

3acbfd2b452e moved the list of things that normalize to a `?` placeholder
into their own function, but there was one more instance of that list to
DRY up.  This commit does that.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

---------

Signed-off-by: Patrick Reynolds <patrick@piki.org>