Fix Insights "too many queries logging" (#1827)
When we moved to the two-tiered query limiter I accidentally left in the
Insights#MaxQueriesPerInterval field, which was not being set and
defaulting to 0. The only place this field was being used was in
logging. Since it defaults to zero the "too many queries" log was
always: "Too many queries: reached limit of 0", which is wrong.

Added limiter#intervalAllotment to record the original number of
messages allowed in this interval so that we can accurately log the
number of messages that were permitted when the limit is exceeded.

Signed-off-by: Rafer Hazen <rafer@ralua.com>