Set QueryStatsBundle#TabletType (#2250)
Add TabletType to the QueryPatternKey and set the value in the
QueryStatsBundle proto. In theory a single vtgate will always have the
same tablet type, but this field is a part of LogStats so technically
it's possible to have different tablet types in the smame vtgate. If
that happens, insights will split aggregates based on the tablet type.

TableType is already set in the Query proto.

Signed-off-by: Rafer Hazen <rafer@ralua.com>