For Insights, always alphabetize `insert` columns (#2544)
* Always alphabetize INSERT columns

Instead of trying to detect when a workload has its INSERT columns in
different but equivalent orders, just always alphabetize them.  The
detection code wasn't working terribly well in pratice: for example,
Barstool's social-data database had 15,000 patterns that should've been
just four.  It takes too long to detect the problem, and the detection is
forgotten whenever a vtgate is rolled or restarted.

So just alphabetize every INSERT...VALUES statement.

This commit also fixes a bug: we no longer alphabetize INSERT...SELECT
statements, because the order of the INSERT columns needs to match the
order of the SELECT columns.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* More tests

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Simplify INSERT column-order normalization

No need to walk the tree twice or do all the type checks.  Instead, use
the fields of the sqlparser.Insert structure directly.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Sort column names on a copy of sqlparser.Insert

We get a the already-parsed SQL statement in LogStats.AST, as of #1796.
Although the LogStats object is only transmitted to the Insights thread
after execution is complete, there's a small chance (now or in the future)
that the thread that generated the LogStats doesn't want us messing with
its AST.  Let's treat the AST as immutable.

That means we need to make a copy of the sqlparser.Insert node before we
alphabetize its columns.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Don't alphabetize single-column lists

No need, as they're already alphabetized.  HT @rafer

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Fix up e2e test expectations

The `-insights_reorder_threshold` flag is deprecated, so the e2e test
needs to expect it as such.

The right way to deprecate a flag appears to be `fs.MarkDeprecated`, but
we can't do that until Insights uses pflag (instead of flag), which is not
worth cramming into this PR.

Signed-off-by: Patrick Reynolds <patrick@piki.org>

* Re-add defaults for e2e test

Even deprecated flags expect to match a default value

Signed-off-by: Patrick Reynolds <patrick@piki.org>

---------

Signed-off-by: Patrick Reynolds <patrick@piki.org>