boost: Support partial materialization for functional columns (#2777)
* boost: Improve upquery generation to include proper aliases

We were dropping aliases before which would break upqueries for derived
tables, since we would lose the names. With the changes here we ensure
that we always keep the names.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

* boost: Support partial materialization for functional columns

This is a bit more involved. When we are able to use a fast path mid
flow upquery, we can directly use the same functional expression when
doing the upquery. In those cases we can still partially materialize.

This doesn't work though if we can't use the fast path. If we have to
fall back to a longer flow for replays, we also have to fall back to
full materialization. This is because we can't invert the functional
expression. There are maybe cases where this would be possible, but it's
not a general solution since not all functions are reversible.

In those cases, we track if we have a functional dependency and if we
fall back to a non-fast replay path, we fall back to full
materialization. This has to happen for the whole dependency chain up
the graph and not just the current node which we guarantee here.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>

---------

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>